+++ tests/fuzz/fuzz.c
@@ -5,6 +5,7 @@
 #include <libpldm/fru.h>
 #include <libpldm/pdr.h>
 #include <libpldm/platform.h>
+#include <libpldm/rde.h>
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
@@ -303,7 +304,30 @@ static int
     return 0;
 }
 
+static int fuzz_decode_pldm_rde_negotiate_redfish_parameters_req(
+    const struct pldm_msg* msg, size_t payload_length)
+{
+    struct pldm_rde_negotiate_redfish_parameters_req req;
+
+    decode_pldm_rde_negotiate_redfish_parameters_req(msg, payload_length, &req);
+
+    return 0;
+}
+
+static int fuzz_decode_pldm_rde_negotiate_redfish_parameters_resp(
+    const struct pldm_msg* msg, size_t payload_length)
+{
+    struct pldm_rde_negotiate_redfish_parameters_resp resp;
+
+    decode_pldm_rde_negotiate_redfish_parameters_resp(msg, payload_length,
+                                                      &resp);
+
+    return 0;
+}
+
 static int (*const decode_pldm_msg_tests[])(const struct pldm_msg*, size_t) = {
+    fuzz_decode_pldm_rde_negotiate_redfish_parameters_req,
+    fuzz_decode_pldm_rde_negotiate_redfish_parameters_resp,
     fuzz_decode_pldm_base_get_tid_resp,
     fuzz_decode_pldm_base_get_pldm_types_resp,
     fuzz_decode_pldm_platform_set_numeric_sensor_enable_resp,
@@ -460,8 +484,81 @@ static int fuzz_encode_pldm_file_df_heartbeat_resp(struct pldm_msg* msg,
     return 0;
 }
 
+static int fuzz_encode_pldm_rde_negotiate_redfish_parameters_req(
+    struct pldm_msg* msg, size_t payload_length, const uint8_t* data,
+    size_t size)
+{
+    struct pldm_rde_negotiate_redfish_parameters_req req;
+    PLDM_MSGBUF_RO_DEFINE_P(buf);
+    uint8_t instance_id;
+    int rc;
+
+    rc = pldm_msgbuf_init_errno(buf, 0, data, size);
+    if (rc)
+    {
+        return -1;
+    }
+
+    pldm_msgbuf_extract(buf, instance_id);
+    pldm_msgbuf_extract(buf, req.mc_concurrency_support);
+    pldm_msgbuf_extract(buf, req.mc_feature_support.value);
+
+    rc = pldm_msgbuf_complete(buf);
+    if (rc)
+    {
+        return -1;
+    }
+
+    encode_pldm_rde_negotiate_redfish_parameters_req(instance_id, &req, msg,
+                                                     &payload_length);
+
+    return 0;
+}
+
+static int fuzz_encode_pldm_rde_negotiate_redfish_parameters_resp(
+    struct pldm_msg* msg, size_t payload_length, const uint8_t* data,
+    size_t size)
+{
+    struct pldm_rde_negotiate_redfish_parameters_resp resp = {0};
+    PLDM_MSGBUF_RO_DEFINE_P(buf);
+    const uint8_t* name;
+    size_t name_len;
+    uint8_t instance_id;
+    int rc;
+
+    rc = pldm_msgbuf_init_errno(buf, 0, data, size);
+    if (rc)
+    {
+        return -1;
+    }
+
+    pldm_msgbuf_extract(buf, instance_id);
+    pldm_msgbuf_extract(buf, resp.completion_code);
+    pldm_msgbuf_extract(buf, resp.device_concurrency_support);
+    pldm_msgbuf_extract(buf, resp.device_capabilities_flags.byte);
+    pldm_msgbuf_extract(buf, resp.device_feature_support.value);
+    pldm_msgbuf_extract(buf, resp.device_configuration_signature);
+    pldm_msgbuf_extract(buf, resp.provider_name_format);
+    pldm_msgbuf_span_remaining(buf, (const void**)&name, &name_len);
+
+    rc = pldm_msgbuf_complete(buf);
+    if (rc)
+    {
+        return -1;
+    }
+
+    resp.provider_name.ptr = name;
+    resp.provider_name.length = name_len;
+    encode_pldm_rde_negotiate_redfish_parameters_resp(instance_id, &resp, msg,
+                                                      &payload_length);
+
+    return 0;
+}
+
 static int (*const encode_pldm_msg_tests[])(struct pldm_msg*, size_t,
                                             const uint8_t*, size_t) = {
+    fuzz_encode_pldm_rde_negotiate_redfish_parameters_req,
+    fuzz_encode_pldm_rde_negotiate_redfish_parameters_resp,
     fuzz_encode_pldm_base_get_tid_resp,
     fuzz_encode_pldm_base_get_pldm_types_resp,
     fuzz_encode_pldm_platform_set_numeric_sensor_enable_req,