+++ CHANGELOG.md
@@ -19,6 +19,12 @@ Change categories:
 
 ### Added
 
+- rde: Add NegotiateRedfishParameters command
+  - Add `encode_pldm_rde_negotiate_redfish_parameters_req()`
+  - Add `decode_pldm_rde_negotiate_redfish_parameters_req()`
+  - Add `encode_pldm_rde_negotiate_redfish_parameters_resp()`
+  - Add `decode_pldm_rde_negotiate_redfish_parameters_resp()`
+
 ### Changed
 
 - doxygen: Enable warnings as errors